ipa: ipu3: af: Auto focus for dw9719 Surface Go2 VCM
Since VCM for surface Go 2 (dw9719) had been successfully
driven, this Af module can be used to control the VCM and
determine the focus value based on the IPU3 AF state.

The variance of each focus step is determined and a greedy
approah is used to find the maximum variance of the AF
state and a appropriate focus value.

Changes since v2:
1. Add grid configuration interface.
2. Add AF accelerator configuration.
3. Move default focus area to the center of the image.

Signed-off-by: Kate Hsuan <hpa@redhat.com>