ipa: ipu3: af: Auto focus for dw9719 Surface Go2 VCM
Since VCM for surface Go 2 (dw9719) had been successfully
driven, this Af module can be used to control the VCM and
determine the focus value based on the IPU3 AF state.

The variance of each focus step is determined and a greedy
approah is used to find the maximum variance of the AF
state and a appropriate focus value.

Changes in v2:
1. Add grid configuration interface.
2. Add AF accelerator configuration.
3. Move default focus area to the center of the image.

Changes in v3:
1. Squash Daniel's commit- Remove v4l2 interaction from AF
   algorithm.
2. Fix greedy AF algorithm since V4l2 interface had been
   removed.
3. Simplify default grid config for AF.
4. AF start area is determined by bdsOutputSize.

Changes in v4:
In v4, it significant improves the AF scan time using a two pass
scaning method and the AF scan will be terminated when it finds
a negative gradient.
1. Introduce 2 pass AF scan (coarse and fine scan) to increase
   the AF success rate.
2. The low pass filter convolution results are used to determine
   a coarse AF result.
3. The high pass convolution result is used to find a good lens
   step in a given AF range determined in pass 1.

Signed-off-by: Kate Hsuan <hpa@redhat.com>