ipa: ipu3: af: AfMode and LensPosition control implementation
AfMode controls include manual, auto, and continuous modes. For auto
and continuous modes, both of the algorithms try to keep the image
focused. The difference between them is that continuous mode has higher
sensitivity to the change of image variance than auto mode. So, the AF
is frequently triggered in continuous mode. The user-defined lens
position can only be set in manual mode.

Signed-off-by: Kate Hsuan <hpa@redhat.com>